+++ packages/thumbprint-react/CHANGELOG.md
@@ -4,7 +4,8 @@
 
 ### Added
 
--   [Minor] Pass `event` object as argument to `onChange` prop of the `Radio` and `Checkbox` components.
+-   [Minor] Pass `event` object as argument to `onChange` prop of the `Radio` and `Checkbox` components
+-   [Minor] Add `onKeyPress` prop to the `Radio` and `Checkbox` components
 
 ## 14.10.0 - 2022-05-09